fix warning

+++ src/node/sql_list_node_test.cc
@@ -33,25 +33,25 @@ class SqlListNodeTest : public ::testing::Test {
 TEST_F(SqlListNodeTest, PushFrontTest) {
     SQLNodeList *pList = manager_->MakeNodeList();
 
-    ASSERT_EQ(0, (int)(pList->GetSize()));
+    ASSERT_EQ(0u, (pList->GetSize()));
 
     pList->PushFront(manager_->MakeLinkedNode(manager_->MakeConstNode(1)));
     pList->PushFront(manager_->MakeLinkedNode(manager_->MakeConstNode(2)));
     pList->PushFront(manager_->MakeLinkedNode(manager_->MakeConstNode(3)));
 
-    ASSERT_EQ(3, pList->GetSize());
+    ASSERT_EQ(3u, pList->GetSize());
     std::cout << *pList << std::endl;
 }
 
 TEST_F(SqlListNodeTest, AppendNodeListTest) {
     SQLNodeList *pList = manager_->MakeNodeList();
     pList->PushFront(manager_->MakeLinkedNode(manager_->MakeConstNode(1)));
-    ASSERT_EQ(1, pList->GetSize());
+    ASSERT_EQ(1u, pList->GetSize());
 
     SQLNodeList *pList2 = manager_->MakeNodeList();
     pList2->PushFront(manager_->MakeLinkedNode(manager_->MakeConstNode(2)));
     pList2->PushFront(manager_->MakeLinkedNode(manager_->MakeConstNode(3)));
-    ASSERT_EQ(2, pList2->GetSize());
+    ASSERT_EQ(2u, pList2->GetSize());
 
     pList->AppendNodeList(pList2);
     std::cout << *pList << std::endl;
