+++ src/sdk/dbms_sdk_test.cc
@@ -57,7 +57,7 @@ TEST_F(DBMSSdkTest, DatabasesAPITest) {
         dbms_sdk->GetDatabases(names, status);
         std::cout << status.msg << std::endl;
         ASSERT_EQ(0, static_cast<int>(status.code));
-        ASSERT_EQ(0, names.size());
+        ASSERT_EQ(0u, names.size());
     }
 
     // create database db1
@@ -93,7 +93,7 @@ TEST_F(DBMSSdkTest, DatabasesAPITest) {
         Status status;
         dbms_sdk->GetDatabases(names, status);
         ASSERT_EQ(0, static_cast<int>(status.code));
-        ASSERT_EQ(3, names.size());
+        ASSERT_EQ(3u, names.size());
     }
 
     {
@@ -111,7 +111,7 @@ TEST_F(DBMSSdkTest, DatabasesAPITest) {
         db.name = "db_not_exist";
         Status status;
         ASSERT_EQ(false, dbms_sdk->IsExistDatabase(db, status));
-        ASSERT_EQ(0, status.code);
+        ASSERT_EQ(0, static_cast<int>(status.code));
     }
 
     delete tablet;
@@ -279,7 +279,7 @@ TEST_F(DBMSSdkTest, TableAPITest) {
 
         dbms_sdk->GetTables(db, names, status);
         ASSERT_EQ(0, static_cast<int>(status.code));
-        ASSERT_EQ(3, names.size());
+        ASSERT_EQ(3u, names.size());
     }
 
     {
@@ -298,7 +298,7 @@ TEST_F(DBMSSdkTest, TableAPITest) {
         Status status;
         dbms_sdk->GetTables(db, names, status);
         ASSERT_EQ(0, static_cast<int>(status.code));
-        ASSERT_EQ(0, names.size());
+        ASSERT_EQ(0u, names.size());
     }
     delete dbms;
     delete dbms_sdk;
